[desktop] Reuse react components from the web app
