+++ apps/desktop/next.config.js
@@ -1,10 +1,16 @@
 /** @type {import('next').NextConfig} */
 
+const { addPackagesPathToSwcLoader } = require('../../packages/config');
+
 const nextConfig = {
   reactStrictMode: true,
   swcMinify: true,
   images: {
     unoptimized: true,
+  },
+  webpack(config) {
+    config.module.rules.forEach(addPackagesPathToSwcLoader);
+    return config;
   },
 };
