+++ apps/desktop/src/components/Pages/Dashboard/index.tsx
@@ -2,12 +2,16 @@ import React from 'react';
 import styles from './index.module.scss';
 import Title from './Title';
 import Header from './Header';
+import { List } from '../..';
 
 export default function Dashboard() {
   return (
     <div className={styles.container}>
       <Title />
       <Header />
+      <List>
+        <li>Feed</li>
+      </List>
     </div>
   );
 }
