+++ apps/desktop/src/components/index.tsx
@@ -1,0 +1,3 @@
+import List from '../../../../packages/ui/List';
+
+export { List };
